update on the settings menu, keys for usage implemented
